Add factory pattern and basic post module
We want to be able to use this setup package in scripts, and also for it
to be maintainable, testable etc.

A Factory pattern seems a good fit here - it allows for a single static
method that can be called to get the setup class, but after that you can
use instance methods in the standard way, rather than littering the
whole package with static methods (which then makes e.g. dependency
injection much harder, if not impossible).

This commit also adds a basic Post module, for generating a post (with
Faker providing the title and content if none is specified), to
illustrate how the Factory pattern will work in practice.