(refactor) Make Setup and array-like object
This means we can pass any number of modules to Setup in an array, and
then drynamically assign them to the $modules array property, but still
access them as if they were individual properties using the `__get()`
method.

In turn, this means that when we add new modules, we don't have to
modify the Setup class, only the Factory class, so this makes future
changes easier to implement.