Add Plugin module
This currently only has one method (activating a plugin). It uses WP CLI
under the hood for this, because the function within the core API
requires a path to a plugin file, rather than just a plugin slug, which
would be much more fiddly to implement when creating a setup script.